Move position and record structs to lib.rs
