--- commit_log/src/position.rs
@@ -1,6 +1,0 @@
-pub enum Position {
-    /// The first entry available.
-    Horizon,
-
-    Offset(usize),
-}
