--- commit_log/src/record.rs
@@ -1,7 +1,0 @@
-pub struct Record {
-    /// The current offset within current segment.
-    pub current_offset: usize,
-
-    /// Index to current segment.
-    pub segment_index: usize,
-}
